[GTK4] Migrate to API version 6.0 of webkit and jsc
With the most recent GTK4s the API version is now 6.0 for webkit/jsc
and a lot of [deprecated items have been
removed](https://webkitgtk.org/reference/webkitgtk/2.50.1/migrating-to-webkitgtk-6.0.html#stop-using-deprecated-apis).

This commit brings some of the key implementations to the new webkit
API when using GTK4.

To make it easier to ensure I don't regress anything I had working,
I marked all known tests that need attention for GTK4 with
`@Tag("gtk4-todo")`. This allows running the test suite with exclude
tag of `gtk4-todo` and have a fast and clean run.

This includes a g_critical call when webkitgtk.c cannot dynamically
load a called method. When a method cannot be loaded it fairly silently
fails (normally returns 0). With the g_critical it makes it easier to
spot the problems.

This includes a change to the looping while waiting for a response
from the async script to make timeout more accurate to the default
10s.

TODOs:

- [ ] I plan to split this PR up into a few component parts
- [ ] Cookies
- [ ] Arrays of items (gtk4_convertToJava missing jsc_value_is_object case)
- [ ] a bunch of OpenWindowListener cases (see `@Tag("gtk4-todo")` tests)
- [ ] probably more
- [ ] check which version of API we can use as a reasonable minimum, in
      particular do we need to deal with API version 5 (version 4 was
      the version of the API that aligned with GTK3)
- [ ] Remove `-DexcludedGroups=gtk4-todo` from GHA workflow

Fixes https://github.com/eclipse-platform/eclipse.platform.swt/issues/2667