Skip build and test on OSes that have no changes
Many of the PRs on SWT are OS specific, this change only runs the build
on OSes if there are changes that can affect that OS.

This is going to be more important with #2714 because potentially a number
of new jobs will be run per PR and avoiding some of the extra time
waiting will be a benefit to all.

The minor negative of skipping tests conditionally is that the **Test
Results** summary comment will report less runs that base commit.

Part of https://github.com/eclipse-platform/eclipse.platform.swt/issues/2714