throw no more handles if a timer could not be created
At the moment the code silently fails on calling timerExec if there
are no handles left or it fails for other reasons.

Relates to https://github.com/eclipse-platform/eclipse.platform.swt/issues/2806
where we can see that we are running out of handles, but the timerExec
does not report a failure.